fix: commonjs optional libs bug workaround
