+++ packages/react-native-sdk/src/hooks/internal/useCallMediaStreamCleanup.ts
@@ -26,7 +26,7 @@ export const useCallMediaStreamCleanup = () => {
         )
       ) {
         // we cleanup media stream only if call is not joined or joining
-        // @ts-expect-error Due to DOM typing incompatible with RN
+        // @ts-ignore Due to DOM typing incompatible with RN
         disposeOfMediaStream(mediaStream);
       }
     };
