+++ packages/react-native-sdk/src/utils/index.ts
@@ -29,9 +29,11 @@ export const generateParticipantTitle = (memberUserId: string) => {
 // Utility to get initials of a name
 export const getInitialsOfName = (name: string) => {
   const names = name.split(' ');
-  let initials = names[0].substring(0, 1).toUpperCase();
+  let initials = (names[0] as string).substring(0, 1).toUpperCase();
   if (names.length > 1) {
-    initials += names[names.length - 1].substring(0, 1).toUpperCase();
+    initials += (names[names.length - 1] as string)
+      .substring(0, 1)
+      .toUpperCase();
   }
   return initials;
 };
