+++ packages/react-native-sdk/src/utils/internal/optionallibs/index.ts
@@ -1,0 +1,2 @@
+export * from './gestureHandler';
+export * from './reanimated';
